enable extraction of extra 2D data
In the original code, there is a problem with the extraction of extra 2D variables (3D are fine). I tried to change the code to enable 2D and 2D variables in one optionbut in the end failed in the allocation part (didn't put too much time into it). Therefore I decided to add an additional option `-xtrac2D` to extract 2D variables in the same way it is done for the 3D fields. To my understanding, using `-xtra2D` one could eliminate the `-ssh` and `-mxl` options.